Organized scripts and prefabs better.
Added a couple of Behaviour scripts that use interfaces.
Modified the FoV script to be able to have line of sight be interrupted by objects.
Created a simple Barrel prefab that is able to interrupt player movement and line of sight.
Cleaned up some code.
Changed the GraphicsSettings so that the sprites get rendered in order based on the value of they y axis.